apiimage/requirements.txt
fastapi
uvicorn
pymongo
requests
Pillow
google-generativeai